// app/types/peoplesection1.ts
export interface Section1Settings {
    id: number;
    background_color: string;
    badge_text: string;
    badge_text_color: string;
    title: string;
    title_color: string;
    description: string;
    description_color: string;
    bullets: string[];
    bullet_color: string;
    button_text: string;
    button_background_color: string;
    button_text_color: string;
    image_url: string;
    image_alt: string;
    created_at: string;
    updated_at: string;
  }